REVISION Target Lock
when u press q, you lock on to target closest to you in range of 1300 units. when u press q while you are locked on to a target, you remove target. when you kill enemy while you have lock on him, lock changes to next enemy closest to the player. please revise unik